move "pkg/plumbing" to "subcmd"
