+++ subcmd/spark/dialer.go
