+++ subcmd/watch/encoding_test.go
