+++ subcmd/watch/server.go
