+++ subcmd/watch/server_test.go
